refactor: 장치 목록 조회를 송출 API 와 같은 Swagger 태그로 묶음
프론트 송출 화면이 장치 목록과 송출을 함께 쓰므로 한 그룹에서 찾을 수
있도록 컨트롤러를 합친다. Swagger 태그 6개 → 4개.

- SpeakerBroadcastController 를 SpeakerController 로 병합 (전광판도 동일)
- 태그 구성: 도메인별 「장치·송출·이력」 과 「프리셋」
