+++ docs/specs/2026-08-25-broadcast-api-design.md
@@ -98,7 +98,7 @@ SpeakerPresetResponse(id, title, message, @JsonUnwrapped BaseResponse)
 | 이력을 **목록/상세로 분리** | `204` 로 간 이상 프론트가 대상별 실패 사유를 볼 경로가 이력뿐이다. 목록에 매번 `results` 를 싣지 않기 위해 요약만 내리고, 상세에서 대상별 결과를 준다 |
 | 이력 경로를 **`GET /speakers/broadcasts`** 로 (송출과 같은 경로) | 같은 컬렉션에 `POST` 는 송출, `GET` 은 이력. 프론트 입장에서 짝이 명확하다 |
 | `BroadcastResult` 에 **`siteId` · `siteName` 포함** | 이력 상세에서 실패한 장치가 어느 현장인지 즉시 보여줄 수 있다 |
-| 프리셋을 `/speakers/presets` 가 아니라 **`/speaker-presets`** 로 분리 | 후속에 `GET /speakers/{id}` 가 붙을 때 경로 변수와 섞이지 않는다. Swagger Tag 도 장치/송출/프리셋으로 나뉘어 프론트 화면 구성과 1:1 로 맞는다 |
+| 프리셋을 `/speakers/presets` 가 아니라 **`/speaker-presets`** 로 분리 | 후속에 `GET /speakers/{id}` 가 붙을 때 경로 변수와 섞이지 않는다. Swagger Tag 는 도메인별로 **장치·송출·이력** 과 **프리셋** 둘로만 나뉜다 — 프론트에서 송출 화면은 장치 목록과 송출을 함께 쓰므로 한 그룹에 두는 편이 찾기 쉽다 |
 | 프리셋 저장소를 **스피커/전광판 분리** | 업체 스펙 확정 후 붙일 옵션 필드(음색·볼륨 vs 스크롤·색상)가 서로 겹치지 않는다. 공유 테이블로 두면 nullable 범벅이 된다 |
 | 송출 대상을 **장치 ID + 현장 ID 둘 다** 허용 | 개별 선택(#26)과 현장 일괄(#28)이 둘 다 요구사항이다. 현장 일괄을 프론트가 목록 조회 후 전체 선택으로 흉내내면 장치가 늘 때마다 요청이 커진다 |
 | 장치 목록 조회는 **비페이징** | 기존 `CctvController.getAll` 과 같은 형태. 한 현장의 장치 수가 페이징이 필요할 규모가 아니다 |
